v0.6更新，添加玩家2，添加了玩家与地图的碰撞箱

+++ Wall.cpp
@@ -2,7 +2,7 @@
 #include "Manager.h"
 using namespace std;
 
-Wall::Wall(double x, double y, int style,double weight)
+Wall::Wall(double x, double y, int style,int weight)
 {
 	this->x = x;
 	this->y = y;
